+++ game/factions/faction.py
@@ -33,6 +33,7 @@ from game.dcs.groundunittype import GroundUnitType
 from game.dcs.shipunittype import ShipUnitType
 from game.dcs.unittype import UnitType
 from pydcs_extensions.f16i_idf.f16i_idf import inject_F16I
+from pydcs_extensions.fa18efg.fa18efg import inject_FA18EFG
 
 if TYPE_CHECKING:
     from game.theater.start_generator import ModSettings
@@ -430,6 +431,8 @@ class Faction:
             self.remove_vehicle("SAM SA-14 Strela-3 manpad")
             self.remove_vehicle("SAM SA-24 Igla-S manpad")
             self.remove_vehicle("Polyana-D4M1 C2 node")
+        if mod_settings.fa_18efg:
+            inject_FA18EFG()
 
     def remove_aircraft(self, name: str) -> None:
         for i in self.aircrafts: