+++ game/theater/start_generator.py
@@ -56,6 +56,7 @@ class ModSettings:
     f4b_phantom: bool = False
     f15d_baz: bool = False
     f_16_idf: bool = False
+    fa_18efg: bool = False
     f22_raptor: bool = False
     f84g_thunderjet: bool = False
     f100_supersabre: bool = False