+++ qt_ui/main.py
@@ -323,6 +323,7 @@ def create_game(
             a4_skyhawk=False,
             a6a_intruder=False,
             f4b_phantom=False,
+            fa_18efg=False,
             f22_raptor=False,
             f84g_thunderjet=False,
             f100_supersabre=False,